requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
pandas==2.1.3
openpyxl==3.1.2
django-rq==2.8.1
rq==1.15.1
redis==5.0.1
python-decouple==3.8
django-cors-headers==4.3.1
python-dateutil==2.8.2
requests==2.31.0